Add files via upload
Added PPT file , source audio file in WAV Format( please do not change audio format) , the main code is in .ipynb file that is runned on google colab.
We two more files that is epochs file which is used to store the epcohs for both LSTM and CNN model in both domains. The epochs file used as input for creating bar charts to observe the performance of the models. We have Mr. Bean clip that illustrate when Audio Classification is useful in life situations.